Rename package jsonobject to json
