+++ ledger/model/contract_execution_result.go
@@ -3,14 +3,14 @@ package model
 import (
 	"fmt"
 
-	"github.com/scalar-labs/scalardl-go-client-sdk/v3/jsonobject"
+	"github.com/scalar-labs/scalardl-go-client-sdk/v3/json"
 	"github.com/scalar-labs/scalardl-go-client-sdk/v3/ledger/asset"
 )
 
 // ContractExecutionResult defines the result of a contract execution.
 // It contains the result of the contract execution along with a list of asset proofs from Ledger and Auditor.
 type ContractExecutionResult struct {
-	Result        jsonobject.JSONObject
+	Result        json.JSONObject
 	Proofs        []asset.Proof
 	AuditorProofs []asset.Proof
 }
@@ -18,8 +18,8 @@ type ContractExecutionResult struct {
 // Equal checks if two contract execution results have the same values.
 func (r ContractExecutionResult) Equal(another ContractExecutionResult) (equal bool) {
 	var (
-		myResult      jsonobject.JSONObject = r.Result
-		anotherResult jsonobject.JSONObject = another.Result
+		myResult      json.JSONObject = r.Result
+		anotherResult json.JSONObject = another.Result
 	)
 
 	if (myResult == nil && anotherResult != nil) || (myResult != nil && anotherResult == nil) {
